refactor: LibraryDetailViewController split of asset picker
PlantAssetsPickerViewController now takes on the role of selecting
images. A delegate was made such that the parent view controller
can run a function when the images have been selected.
